requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
langchain
langchain-community
sentence-transformers
chromadb
openai
python-dotenv
pydantic
huggingface-hub
transformers
torch
accelerate
bitsandbytes
ollama
requests
python-dotenv
